+++ Libplanet.Tests/Action/ActionEvaluatorTest.cs
@@ -79,13 +79,14 @@ namespace Libplanet.Tests.Action
                 transactions: txs
             );
             Block<RandomAction> stateRootBlock =
-                noStateRootBlock.Evaluate(GenesisMiner, null, stateStore);
+                noStateRootBlock.Evaluate(GenesisMiner, null, _ => true, stateStore);
             var actionEvaluator =
                 new ActionEvaluator<RandomAction>(
                     policyBlockAction: null,
                     blockChainStates: NullChainStates<RandomAction>.Instance,
                     trieGetter: null,
-                    genesisHash: null);
+                    genesisHash: null,
+                    nativeTokenPredicate: _ => true);
             var generatedRandomNumbers = new List<int>();
 
             AssertPreEvaluationBlocksEqual(stateRootBlock, noStateRootBlock);
@@ -281,7 +282,9 @@ namespace Libplanet.Tests.Action
                 policyBlockAction: null,
                 blockChainStates: NullChainStates<DumbAction>.Instance,
                 trieGetter: null,
-                genesisHash: null);
+                genesisHash: null,
+                nativeTokenPredicate: _ => true
+            );
             IAccountStateDelta previousStates = genesis.ProtocolVersion > 0
                 ? new AccountStateDeltaImpl(
                     ActionEvaluator<DumbAction>.NullAccountStateGetter,
@@ -584,7 +587,8 @@ namespace Libplanet.Tests.Action
                 policyBlockAction: null,
                 blockChainStates: NullChainStates<DumbAction>.Instance,
                 trieGetter: null,
-                genesisHash: tx.GenesisHash);
+                genesisHash: tx.GenesisHash,
+                nativeTokenPredicate: _ => true);
 
             foreach (bool rehearsal in new[] { false, true })
             {
@@ -712,7 +716,9 @@ namespace Libplanet.Tests.Action
                 policyBlockAction: null,
                 blockChainStates: NullChainStates<ThrowException>.Instance,
                 trieGetter: null,
-                genesisHash: tx.GenesisHash);
+                genesisHash: tx.GenesisHash,
+                nativeTokenPredicate: _ => true
+            );
             var block = new BlockContent<ThrowException>
             {
                 Index = 123,
@@ -761,7 +767,9 @@ namespace Libplanet.Tests.Action
                 actions: txA.Actions.ToImmutableArray<IAction>(),
                 rehearsal: rehearsal,
                 previousBlockStatesTrie: fx.GetTrie(blockA.PreviousHash),
-                blockAction: false).ToArray();
+                blockAction: false,
+                nativeTokenPredicate: _ => true
+            ).ToArray();
 
             Assert.Equal(evalsA.Length, deltaA.Count - 1);
             for (int i = 0; i < evalsA.Length; i++)
@@ -811,7 +819,9 @@ namespace Libplanet.Tests.Action
                 actions: txB.Actions.ToImmutableArray<IAction>(),
                 rehearsal: rehearsal,
                 previousBlockStatesTrie: fx.GetTrie(blockB.PreviousHash),
-                blockAction: false).ToArray();
+                blockAction: false,
+                nativeTokenPredicate: _ => true
+            ).ToArray();
 
             Assert.Equal(evalsB.Length, deltaB.Count - 1);