+++ Libplanet/Action/ActionContext.cs
@@ -1,5 +1,7 @@
+using System;
 using System.Diagnostics.Contracts;
 using System.Security.Cryptography;
+using Libplanet.Assets;
 using Libplanet.Blocks;
 using Libplanet.Store.Trie;
 using Libplanet.Tx;
@@ -10,7 +12,7 @@ namespace Libplanet.Action
     {
         private readonly int _randomSeed;
         private readonly ITrie? _previousBlockStatesTrie;
-
+        private readonly Predicate<Currency>? _nativeTokenPredicate;
         private HashDigest<SHA256>? _previousStateRootHash;
 
         public ActionContext(
@@ -23,8 +25,8 @@ namespace Libplanet.Action
             int randomSeed,
             bool rehearsal = false,
             ITrie? previousBlockStatesTrie = null,
-            bool blockAction = false
-        )
+            bool blockAction = false,
+            Predicate<Currency>? nativeTokenPredicate = null)
         {
             GenesisHash = genesisHash;
             Signer = signer;
@@ -37,6 +39,7 @@ namespace Libplanet.Action
             _randomSeed = randomSeed;
             _previousBlockStatesTrie = previousBlockStatesTrie;
             BlockAction = blockAction;
+            _nativeTokenPredicate = nativeTokenPredicate;
         }
 
         public BlockHash? GenesisHash { get; }
@@ -68,6 +71,9 @@ namespace Libplanet.Action
 
         public bool BlockAction { get; }
 
+        public bool IsNativeToken(Currency currency) =>
+            _nativeTokenPredicate is { } && _nativeTokenPredicate(currency);
+
         [Pure]
         public IActionContext GetUnconsumedContext() =>
             new ActionContext(
@@ -80,6 +86,7 @@ namespace Libplanet.Action
                 _randomSeed,
                 Rehearsal,
                 _previousBlockStatesTrie,
-                BlockAction);
+                BlockAction,
+                _nativeTokenPredicate);
     }
 }