+++ Libplanet/Action/ActionEvaluator.cs
@@ -30,6 +30,7 @@ namespace Libplanet.Action
         private readonly IAction? _policyBlockAction;
         private readonly IBlockChainStates<T> _blockChainStates;
         private readonly Func<BlockHash, ITrie>? _trieGetter;
+        private readonly Predicate<Currency> _nativeTokenPredicate;
 
         /// <summary>
         /// Creates a new <see cref="ActionEvaluator{T}"/>.
@@ -43,17 +44,22 @@ namespace Libplanet.Action
         /// a provided <see cref="BlockHash"/>.</param>
         /// <param name="genesisHash"> A <see cref="BlockHash"/> value of the genesis block.
         /// </param>
+        /// <param name="nativeTokenPredicate">A predicate function to determine whether
+        /// the specified <see cref="Currency"/> is a native token defined by chain's
+        /// <see cref="Libplanet.Blockchain.Policies.IBlockPolicy{T}.NativeTokens"/> or not.</param>
         public ActionEvaluator(
             IAction? policyBlockAction,
             IBlockChainStates<T> blockChainStates,
             Func<BlockHash, ITrie>? trieGetter,
-            BlockHash? genesisHash)
+            BlockHash? genesisHash,
+            Predicate<Currency> nativeTokenPredicate)
         {
             _logger = Log.ForContext<ActionEvaluator<T>>();
             _policyBlockAction = policyBlockAction;
             _blockChainStates = blockChainStates;
             _trieGetter = trieGetter;
             _genesisHash = genesisHash;
+            _nativeTokenPredicate = nativeTokenPredicate;
         }
 
         /// <summary>
@@ -174,7 +180,8 @@ namespace Libplanet.Action
                 signature: tx.Signature,
                 actions: tx.Actions.Cast<IAction>().ToImmutableList(),
                 rehearsal: true,
-                previousBlockStatesTrie: null);
+                previousBlockStatesTrie: null,
+                nativeTokenPredicate: _ => true);
 
             if (evaluations.Any())
             {
@@ -207,6 +214,9 @@ namespace Libplanet.Action
         /// <param name="signature"><see cref="Transaction{T}"/> signature used to generate random
         /// seeds.</param>
         /// <param name="actions">Actions to evaluate.</param>
+        /// <param name="nativeTokenPredicate">A predicate function to determine whether
+        /// the specified <see cref="Currency"/> is a native token defined by chain's
+        /// <see cref="Libplanet.Blockchain.Policies.IBlockPolicy{T}.NativeTokens"/> or not.</param>
         /// <param name="rehearsal">Pass <c>true</c> if it is intended
         /// to be dry-run (i.e., the returned result will be never used).
         /// The default value is <c>false</c>.</param>
@@ -246,6 +256,7 @@ namespace Libplanet.Action
             Address signer,
             byte[] signature,
             IImmutableList<IAction> actions,
+            Predicate<Currency> nativeTokenPredicate,
             bool rehearsal = false,
             ITrie? previousBlockStatesTrie = null,
             bool blockAction = false,
@@ -263,7 +274,8 @@ namespace Libplanet.Action
                     randomSeed: randomSeed,
                     rehearsal: rehearsal,
                     previousBlockStatesTrie: previousBlockStatesTrie,
-                    blockAction: blockAction);
+                    blockAction: blockAction,
+                    nativeTokenPredicate: nativeTokenPredicate);
             }
 
             byte[] hashedSignature;
@@ -547,7 +559,8 @@ namespace Libplanet.Action
                 signature: tx.Signature,
                 actions: tx.Actions.Cast<IAction>().ToImmutableList(),
                 rehearsal: rehearsal,
-                previousBlockStatesTrie: previousBlockStatesTrie);
+                previousBlockStatesTrie: previousBlockStatesTrie,
+                nativeTokenPredicate: _nativeTokenPredicate);
 
         /// <summary>
         /// Evaluates <see cref="Transaction{T}.Actions"/> of a given
@@ -631,7 +644,9 @@ namespace Libplanet.Action
                 actions: new[] { _policyBlockAction }.ToImmutableList(),
                 rehearsal: false,
                 previousBlockStatesTrie: previousBlockStatesTrie,
-                blockAction: true).Single();
+                blockAction: true,
+                nativeTokenPredicate: _nativeTokenPredicate
+            ).Single();
         }
 
         [Pure]